requirements.txt
git+https://github.com/KurimuzonAkuma/pyrogram.git # 2.2.4 !!!
qrcode==7.4.2
psutil>=6.0.0
ujson>=5.10.0
gitpython==3.1.43
aiogram>=3,<4
pyyaml>=6.0.2
websockets>=15.0.1
colorlog # 6.9.0 